Renames var C to var B
