+++ src/fronts/opinion/variantB/components/LinkCollection.tsx
